Moving cow_counters test to unit-test, from top-level integration test
